feat: rename pack to wordpack
